Encrypt passwords

+++ Blog/View/UserAccountAdminView.swift
@@ -26,6 +26,12 @@ struct UserAccountAdminView: View {
     private var animateTextInternalIdCopied = false
     @State
     private var isConfirmDeletePresented = false
+    @State
+    private var displaySectionUserSecrets = false
+    
+    private var isOwnAccount: Bool {
+        accountManager.loggedInUser?.id == user.id
+    }
     
     private func onPressTextInternalId() {
         UIPasteboard.general.string = user.id
@@ -69,13 +75,28 @@ struct UserAccountAdminView: View {
         case 4: "Superuser"
         default: "Default"
         }
+        let selfId = accountManager.loggedInUser?.id
         
         return HStack {
             Text("Rank: \(title) (\(level))")
             Spacer()
-            buttonPromote
-            buttonDemote
+            if user.permissionLevel != 4 && user.id != selfId {
+                buttonPromote
+                buttonDemote
+            }
         }
+    }
+    
+    private var textInferiorRank: some View {
+        let bool = user.isRankInferiorTo(accountManager.loggedInUser)
+        let binaryWord = bool ? "Yes" : "No"
+        return Text("Inferior Rank: \(binaryWord)")
+    }
+    
+    private var textSuperiorRank: some View {
+        let bool = user.isRankSuperiorTo(accountManager.loggedInUser)
+        let binaryWord = bool ? "Yes" : "No"
+        return Text("Superior Rank: \(binaryWord)")
     }
     
     private var textCommentsCount: some View {
@@ -100,12 +121,30 @@ struct UserAccountAdminView: View {
         return Text("Biography: \"\(user.biography)\"")
     }
     
+    private var textPassword: some View {
+        return HStack {
+            Text("Password: \(user.password)")
+            Spacer()
+            
+            if !isOwnAccount {
+                Button("Log In As") {
+                    accountManager.loggedInUser = user
+                    dismiss()
+                }
+            }
+        }
+    }
+    
     private var buttonDelete: some View {
         func onPress() {
             isConfirmDeletePresented = true
         }
         
         func onConfirm() {
+            if user.isSuperUser() {
+                return
+            }
+            
             isConfirmDeletePresented = false
             modelContext.delete(user)
             try? modelContext.save()
@@ -116,6 +155,7 @@ struct UserAccountAdminView: View {
             onPress()
         }
         .buttonStyle(.bordered)
+        .disabled(user.isSuperUser())
         .tint(.red)
         .foregroundStyle(.red)
         .confirmationDialog("Delete user \"\(user.username)\"? (Cannot be undone)", isPresented: $isConfirmDeletePresented, titleVisibility: .visible, actions: {
@@ -151,7 +191,9 @@ struct UserAccountAdminView: View {
         }
         
         func isDisabled() -> Bool {
-            user.permissionLevel == 3 // Cannot promote past operator
+            user.isRankSuperiorTo(accountManager.loggedInUser) || // Cannot promote a superior user
+            user.permissionLevel == accountManager.loggedInUser?.permissionLevel || // Cannot promote an equal user
+            user.permissionLevel >= 3 // Cannot promote past operator
         }
         
         return Button(action: onPress) {
@@ -184,7 +226,7 @@ struct UserAccountAdminView: View {
                 return
             }
             
-            if !thisUser!.isRankSuperiorTo(thatUser) {
+            if thisUser!.isRankInferiorTo(thatUser) {
                 return
             }
             
@@ -199,7 +241,9 @@ struct UserAccountAdminView: View {
         }
         
         func isDisabled() -> Bool {
-            user.permissionLevel == 0 // Cannot demote past default
+            user.isRankSuperiorTo(accountManager.loggedInUser) || // Cannot demote a superior user
+            user.permissionLevel == 0 || // Cannot demote lower than default
+            user.permissionLevel == 4    // Cannot demote a superuser
         }
         
         return Button(action: onPress) {
@@ -211,19 +255,42 @@ struct UserAccountAdminView: View {
         .disabled(isDisabled())
     }
     
+    @ViewBuilder
+    private var buttonDisplaySecrets: some View {
+        let isSuperUser = accountManager.loggedInUser?.isSuperUser() ?? false
+        
+        if isSuperUser {
+            Button("Display Secrets") {
+                displaySectionUserSecrets.toggle()
+            }
+            .buttonStyle(.bordered)
+            .tint(.yellow)
+        } else {
+            EmptyView()
+        }
+    }
+    
     public var body: some View {
         VStack {
             List {
-                Section("Data") {
+                Section("User Data") {
                     textInternalId
                     textJoined
                     textPermissionLevel
+                    textInferiorRank
+                    textSuperiorRank
                 }
                 
                 Section("User-Configured") {
                     textUsername
                     textDisplayName
                     textBiography
+                }
+                
+                if displaySectionUserSecrets {
+                    Section("User Secrets") {
+                        textPassword
+                    }
                 }
                 
                 Section("Statistics") {
@@ -236,7 +303,9 @@ struct UserAccountAdminView: View {
                         buttonDelete
                         // TODO Force rename
                         // TODO Force clear bio
-                        
+                    }
+                    HStack {
+                        buttonDisplaySecrets
                     }
                 }
             }
